refactor: centralize stakes score types
